代码规范

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/market/friendpromote/FriendPromoteLaunchParam.java
@@ -7,23 +7,23 @@ import lombok.Data;
  */
 @Data
 public class FriendPromoteLaunchParam {
-	/* 好友助力活动Id（商家发起） */
+	/** 好友助力活动Id（商家发起） */
 	private Integer promoteId;
 	
-	/* 领取用户昵称 */
+	/** 领取用户昵称 */
 	private String username;
 	
-	/* 领取用户手机号 */
+	/** 领取用户手机号 */
 	private String mobile;
 	
-	/* 助力活动Id（用户发起） */
+	/** 助力活动Id（用户发起） */
 	private Integer id;
 
-	/* 是否助力成功 */
-	public static final int PROMOTE_STATUS_DEFAULT = -1;
-	private int promoteStatus = PROMOTE_STATUS_DEFAULT;
+	/** 是否助力成功 */
+	public static final Byte PROMOTE_STATUS_DEFAULT = -1;
+	private Byte promoteStatus = PROMOTE_STATUS_DEFAULT;
 	
-	/* 分页信息 */
-	private int currentPage;
-	private int pageRows;
+	/** 分页信息 */
+	private Integer currentPage;
+	private Integer pageRows;
 }
